+++ plugins/ml/src/main/java/greycat/ml/regression/PolynomialNode.java
@@ -65,7 +65,7 @@ public class PolynomialNode extends BaseMLNode implements RegressionNode {
 
     //Override default Abstract node default setters and getters
     @Override
-    public final Node set(String propertyName, byte propertyType, Object propertyValue) {
+    public final Node set(String propertyName, int propertyType, Object propertyValue) {
         if (propertyName.equals(VALUE)) {
             learn(Double.parseDouble(propertyValue.toString()), null);
         } else if (propertyName.equals(PRECISION)) {
